pulling from original

+++ osmdroid-android/src/main/java/org/osmdroid/tileprovider/tilesource/BitmapTileSourceBase.java
@@ -1,26 +1,22 @@
 package org.osmdroid.tileprovider.tilesource;
 
+import android.graphics.Bitmap;
+import android.graphics.BitmapFactory;
+import android.graphics.drawable.Drawable;
+import android.os.Build;
+import android.util.Log;
+
+import org.osmdroid.api.IMapView;
+import org.osmdroid.tileprovider.BitmapPool;
+import org.osmdroid.tileprovider.MapTile;
+import org.osmdroid.tileprovider.ReusableBitmapDrawable;
+import org.osmdroid.tileprovider.util.Counters;
+
 import java.io.File;
 import java.io.InputStream;
 import java.util.Random;
 
-import org.osmdroid.ResourceProxy;
-import org.osmdroid.ResourceProxy.string;
-import org.osmdroid.tileprovider.BitmapPool;
-import org.osmdroid.tileprovider.MapTile;
-import org.osmdroid.tileprovider.ReusableBitmapDrawable;
-import org.osmdroid.tileprovider.constants.OpenStreetMapTileProviderConstants;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
-import android.graphics.drawable.Drawable;
-
-public abstract class BitmapTileSourceBase implements ITileSource,
-		OpenStreetMapTileProviderConstants {
-
-	private static final Logger logger = LoggerFactory.getLogger(BitmapTileSourceBase.class);
+public abstract class BitmapTileSourceBase implements ITileSource {
 
 	private static int globalOrdinal = 0;
 
@@ -28,27 +24,25 @@ public abstract class BitmapTileSourceBase implements ITileSource,
 	private final int mMaximumZoomLevel;
 
 	private final int mOrdinal;
-	protected final String mName;
+	protected String mName;
 	protected final String mImageFilenameEnding;
 	protected final Random random = new Random();
 
 	private final int mTileSizePixels;
 
-	private final string mResourceId;
+	//private final string mResourceId;
 
 	/**
 	 * Constructor
-	 * @param aName a human-friendly name for this tile source
-	 * @param aResourceId resource id used to get the localized name of this tile source
+	 * @param aName a human-friendly name for this tile source. this name is also used on the file system, to keep the characters linux file system friendly
 	 * @param aZoomMinLevel the minimum zoom level this tile source can provide
 	 * @param aZoomMaxLevel the maximum zoom level this tile source can provide
 	 * @param aTileSizePixels the tile size in pixels this tile source provides
 	 * @param aImageFilenameEnding the file name extension used when constructing the filename
 	 */
-	public BitmapTileSourceBase(final String aName, final string aResourceId,
+	public BitmapTileSourceBase(final String aName,
 			final int aZoomMinLevel, final int aZoomMaxLevel, final int aTileSizePixels,
 			final String aImageFilenameEnding) {
-		mResourceId = aResourceId;
 		mOrdinal = globalOrdinal++;
 		mName = aName;
 		mMinimumZoomLevel = aZoomMinLevel;
@@ -90,31 +84,43 @@ public abstract class BitmapTileSourceBase implements ITileSource,
 		return mTileSizePixels;
 	}
 
-	@Override
-	public String localizedName(final ResourceProxy proxy) {
-		return proxy.getString(mResourceId);
-	}
 
 	@Override
-	public Drawable getDrawable(final String aFilePath) {
+	public Drawable getDrawable(final String aFilePath) throws LowMemoryException {
+		//Log.d(IMapView.LOGTAG, aFilePath + " attempting to load bitmap");
 		try {
 			// default implementation will load the file as a bitmap and create
 			// a BitmapDrawable from it
 			BitmapFactory.Options bitmapOptions = new BitmapFactory.Options();
 			BitmapPool.getInstance().applyReusableOptions(bitmapOptions);
-			final Bitmap bitmap = BitmapFactory.decodeFile(aFilePath, bitmapOptions);
+			final Bitmap bitmap;
+			//fix for API 15 see https://github.com/osmdroid/osmdroid/issues/227
+			if (Build.VERSION.SDK_INT == Build.VERSION_CODES.ICE_CREAM_SANDWICH_MR1)
+				bitmap=BitmapFactory.decodeFile(aFilePath);
+			else
+				bitmap = BitmapFactory.decodeFile(aFilePath, bitmapOptions);
 			if (bitmap != null) {
 				return new ReusableBitmapDrawable(bitmap);
 			} else {
-				// if we couldn't load it then it's invalid - delete it
-				try {
-					new File(aFilePath).delete();
-				} catch (final Throwable e) {
-					logger.error("Error deleting invalid file: " + aFilePath, e);
-				}
+				File bmp = new File(aFilePath);
+				if (bmp.exists()) {
+					// if we couldn't load it then it's invalid - delete it
+					Log.d(IMapView.LOGTAG, aFilePath + " is an invalid image file, deleting...");
+					try {
+						new File(aFilePath).delete();
+					} catch (final Throwable e) {
+						Log.e(IMapView.LOGTAG, "Error deleting invalid file: " + aFilePath, e);
+					}
+				} else
+					Log.d(IMapView.LOGTAG, "Request tile: " + aFilePath + " does not exist");
 			}
 		} catch (final OutOfMemoryError e) {
-			logger.error("OutOfMemoryError loading bitmap: " + aFilePath);
+			Log.e(IMapView.LOGTAG,"OutOfMemoryError loading bitmap: " + aFilePath);
+			System.gc();
+			throw new LowMemoryException(e);
+		} catch (final Exception e){
+			Log.e(IMapView.LOGTAG,"Unexpected error loading bitmap: " + aFilePath,e);
+			Counters.tileDownloadErrors++;
 			System.gc();
 		}
 		return null;
@@ -146,14 +152,14 @@ public abstract class BitmapTileSourceBase implements ITileSource,
 				return new ReusableBitmapDrawable(bitmap);
 			}
 		} catch (final OutOfMemoryError e) {
-			logger.error("OutOfMemoryError loading bitmap");
+			Log.e(IMapView.LOGTAG,"OutOfMemoryError loading bitmap");
 			System.gc();
 			throw new LowMemoryException(e);
 		}
 		return null;
 	}
 
-	public final class LowMemoryException extends Exception {
+	public static final class LowMemoryException extends Exception {
 		private static final long serialVersionUID = 146526524087765134L;
 
 		public LowMemoryException(final String pDetailMessage) {
